+++ examples/Basics/user_callable.py
@@ -1,14 +1,13 @@
-## Example for a user defined callable
+## Example of a user-defined callable
 
-# This example is an extenssion of the recommenders example.
-# It implements an End-to-End example, that showes how to use the user defined callable in pure recommenders.
+# This example is an extension of the recommender example.
+# It implements an end-to-end use case demonstrating how to utilize the user-defined callable in pure recommenders.
 
 # This examples assumes some basic familiarity with using BayBE.
 # We refer to [`campaign`](./campaign.md) for a more general and basic example.
 
 ### Necessary imports for this example
 
-import pickle
 from typing import Optional
 
 import numpy as np
@@ -23,36 +22,14 @@ from baybe.recommenders import (
     TwoPhaseMetaRecommender,
 )
 from baybe.searchspace import SearchSpace
-from baybe.surrogates import GaussianProcessSurrogate
 from baybe.targets import NumericalTarget
 from baybe.utils.dataframe import add_fake_results
 
-# Per default the initial recommender chosen is a random recommender.
-INITIAL_RECOMMENDER = RandomRecommender()
-
-### Surrogate models
-
-# This model uses available data to model the objective function as well as the uncertainty.
-# The surrogate model is then used by the acquisition function to make recommendations.
-
-# Per default a Gaussian Process is used
-SURROGATE_MODEL = GaussianProcessSurrogate()
-
-### Acquisition function
-
-# This function looks for points where measurements of the target value could improve the model.
-
-# Note that the qvailability of the acquisition functions might depend on the `batch_size`:
-#   - If `batch_size` is set to 1, all available acquisition functions can be chosen
-#   - If a larger value is chosen, only those that allow batching.
-#       That is, 'q'-variants of the acquisition functions must be chosen.
-
-# The default acquisition function is q-Expected Improvement.
-
-ACQ_FUNCTION = "qEI"
-
-
 ### Callable function
+
+# This function prints all given parameters and pickles the recommender object to test its serializability.
+
+
 def callable_test(
     recommender: Optional[SequentialGreedyRecommender] = None,
     searchspace: Optional[SearchSpace] = None,
@@ -68,43 +45,17 @@ def callable_test(
         batch_size: The number of points to be recommended.
         train_x: Optional training inputs for training a model.
         train_y: Optional training labels for training a model.
-
-    This function prints all giving parameters and pickles the recommender object
-    to test the serializability.
     """
     print("Start callable_test")
-    print(
-        f"""{recommender}\n{searchspace}\nBatch Size:{batch_size}
-    \ntrain x:\n{train_x}\ntrain y:\n{train_y}"""
-    )
-    with open("callableTest.pkl", "wb") as file:
-        pickle.dump(recommender, file)
+    print(f"""\ntrain x:\n{train_x}\ntrain y:\n{train_y}""")
     print("End callable_test")
 
 
-### Other parameters
-
-# Two other boolean hyperparameters can be specified when creating a recommender object.
-# The first one allows the recommendation of points that were already recommended previously.
-# The second one allows the recommendation of points that have already been measured.
-# Per default, they are set to `True`.
-
-ALLOW_REPEATED_RECOMMENDATIONS = True
-ALLOW_RECOMMENDING_ALREADY_MEASURED = True
-
 ### Creating the recommender object
 
-# To create the recommender object, each parameter described above can be specified as follows.
-# Note that they all have default values.
-# Therefore one does not need to specify all of them to create a recommender object.
-
 recommender = TwoPhaseMetaRecommender(
-    initial_recommender=INITIAL_RECOMMENDER,
+    initial_recommender=RandomRecommender(),
     recommender=SequentialGreedyRecommender(
-        surrogate_model=SURROGATE_MODEL,
-        acquisition_function=ACQ_FUNCTION,
-        allow_repeated_recommendations=ALLOW_REPEATED_RECOMMENDATIONS,
-        allow_recommending_already_measured=ALLOW_RECOMMENDING_ALREADY_MEASURED,
         user_callable=callable_test,
     ),
 )
@@ -144,9 +95,8 @@ campaign = Campaign(
     objective=objective,
 )
 
-# This campaign can then be used to get recommendations and add measurements:
-# We use a loop to make sure that the recommend call of the pure recommender is made.
-# The initial recommend call is made in the campaign, and from that point on, the recommend call of the pure recommender will be used to train the recommender based on the previous recommendations.
+# We now run some loops of the recommendation process to demonstrate the use of the callable
+
 for i in range(2):
     recommendation = campaign.recommend(batch_size=3)
     print("\n\nRecommended experiments: ")